refactored pom.xml from dead dependency, added swagger
